requirements.txt
asgiref==3.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.1
django-crispy-forms==2.1
django-randomslugfield==0.3.0
djangorestframework==3.14.0
idna==3.7
pillow==10.2.0
pyTelegramBotAPI==4.19.2
python-slugify==8.0.3
pytz==2024.1
requests==2.32.3
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.4
urllib3==2.2.2
